refactor: update import paths for listing form schema and remove obsolete file
