+++ apps/web/components/merchant/steps/PaymentLimitsStep.tsx
@@ -16,7 +16,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select';
-import type { ListingFormValues } from '@/lib/listing-form-schema';
+import type { ListingFormValues } from '@/lib/schemas/listing/listing-form-schema';
 
 export function PaymentLimitsStep() {
   const form = useFormContext<ListingFormValues>();
